# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZyNum API - Virtual Numbers Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/balance:
    get:
      operationId: getBalance
      tags: [balance]
      summary: Get 5SIM balance
      responses:
        "200":
          description: Balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
        currency:
          type: string
        lowBalanceThreshold:
          type: number
        isLow:
          type: boolean
      required:
        - balance
        - currency
        - lowBalanceThreshold
        - isLow
...